@@ -1,35 +1,40 @@
 {
   "name": "zoomers-collab-timers_frontend",
   "dependencies": {
+    "browserify-zlib": "^0.2.0",
     "cra-append-sw": "^2.7.0",
-    "date-fns": "^2.24.0",
-    "debug": "^4.3.2",
-    "firebase": "^9.0.2",
+    "date-fns": "^2.29.3",
+    "debug": "^4.3.4",
+    "firebase": "^9.12.1",
     "formik": "^2.2.9",
     "nanoid": "^3.1.25",
+    "os-browserify": "^0.3.0",
     "push.js": "^1.0.12",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
+    "react-app-rewired": "^2.2.1",
     "react-collapse": "^5.1.0",
-    "react-dom": "^17.0.2",
-    "react-firebase-hooks": "^3.0.4",
-    "react-ga": "^3.3.0",
+    "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.0.3",
+    "react-ga": "^3.3.1",
     "react-gravatar": "^2.6.3",
     "react-hooks-global-state": "^2.0.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "react-transition-group": "^4.4.2",
-    "sass": "^1.41.1",
-    "serve": "^12.0.1",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
+    "react-scripts": "5.0.1",
+    "react-transition-group": "^4.4.5",
+    "sass": "^1.55.0",
+    "serve": "^14.0.1",
     "socket.io-client": "^4.2.0",
+    "web-vitals": "^2.1.4",
     "winston": "3.2.1",
     "winston-transport-browserconsole": "^1.0.5",
-    "yup": "^0.32.9"
+    "yup": "^0.32.11"
   },
   "scripts": {
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "start": "react-scripts start",
+    "start": "react-app-rewired start",
     "cypress:open": "cypress open",
     "go": "./node_modules/pm2/bin/pm2 start ecosystem.config.js",
     "goSemi": "cd ./../sockets && pm2 start nodemon -- src/server.js && cd ./../frontend && pm2 start react-scripts -- start",
