@@ -1,10 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.scss';
+import { createRoot } from 'react-dom/client';
 import AppRouter from './AppRouter';
 import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render( <AppRouter />, document.getElementById( 'root' ) );
+const container = document.getElementById( 'root' );
+// const container = document.getElementById( 'app' );
+const root = createRoot( container ); // container! if TS
+root.render( <AppRouter /> );
+
+// Old
+// ReactDOM.render( <AppRouter />, document.getElementById( 'root' ) );
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
